+++ backend/bittan/bittan/models/__init__.py
@@ -1,0 +1,4 @@
+from .chapter_event import ChapterEvent
+from .payment import Payment
+from .ticket_type import TicketType
+from .ticket import Ticket
