+++ backend/bittan/bittan/models/chapter_event.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+class ChapterEvent(models.Model):
+	title = models.TextField()
+	description = models.TextField()
+	max_tickets = models.IntegerField()
+	sales_stop_at = models.DateTimeField()
+	ticket_types = models.ManyToManyField('TicketType')
